simple_alphazero/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
numpy>=1.19.5
chess>=1.9.0
pygame>=2.0.1
cairosvg>=2.5.2
pillow>=8.3.1
matplotlib>=3.4.3
tqdm>=4.62.3
